adding before scenario hook to Regression and Sys Tests which checks that auth states are valid before executing scenario + relocating existing bdd hooks to this file and removing from current locations (#41)
